splitting components into their own files, cleaning up comments and logging
